requirements_as_lib.txt
httpx>=0.7
python-frontmatter>=0.4
coloredlogs>=10
python-dotenv>=0.10
